// generated/src/app/api-models/al-letter-event-tickler.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AlLetterEventTickler {
  seqLetterEventId: number;
  letterEventId: string;
  key1: string;
  key2: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}